Connected Utils refactored.
