+++ src/main/java/cl/set/markito/IDebugManager.java
@@ -7,6 +7,7 @@ import java.io.PrintStream;
  */
 
 public interface IDebugManager {
+    public boolean getDebugMode();
     public PrintStream getDebugManagerOutputStream();
     public void setDebugManagerOutputStream( PrintStream output );
     public void clearConsole();
