+++ src/main/java/cl/set/markito/RandomUtils.java
@@ -4,11 +4,7 @@ import java.util.concurrent.ThreadLocalRandom;
 
 public class RandomUtils implements IRandomUtils {
     
-    /**
-     * Generates a random String of size length.
-     * @param size
-     * @return
-     */
+
     public String RandomString(int size) {
         String pool = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890";
         String randomString = "";
@@ -20,12 +16,6 @@ public class RandomUtils implements IRandomUtils {
         return randomString;
     }
 
-    /**
-     * Generates a random integer in range [min, max].
-     * @param min
-     * @param max
-     * @return
-     */
     public int RandomNumber(int min, int max) {
         return ThreadLocalRandom.current().nextInt(min, max + 1);
     }
