+++ src/test/java/cl/set/markito/samples/AndroidTestWikipediaInBrowserStack.java
@@ -8,6 +8,7 @@ import org.junit.jupiter.api.Test;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.remote.DesiredCapabilities;
 
+import cl.set.markito.BrowserStack;
 import cl.set.markito.MarkitoAndroid;
 import io.appium.java_client.MobileBy;
 
@@ -37,9 +38,10 @@ public class AndroidTestWikipediaInBrowserStack extends MarkitoAndroid {
 	}
 
 	private void setDesiredCapabilities(DesiredCapabilities caps) {
+		BrowserStack bs = new BrowserStack();
 		// Set your access credentials
-		caps.setCapability("browserstack.user", getBsUsername());
-		caps.setCapability("browserstack.key", getBsPassword());
+		caps.setCapability("browserstack.user", bs.getBsUsername());
+		caps.setCapability("browserstack.key", bs.getBsPassword());
 		caps.setCapability("browserstack.appium_version", "1.17.0");
 		
 		// Set URL of the application under test
